backend/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
supabase
python-dotenv